+++ TJC.GitExtensions/Branches.cs
@@ -2,18 +2,18 @@ namespace TJC.GitExtensions;
 
 public static partial class GitExtensions
 {
-    public static string GetBranchName(string workingDirectory = ".")
+    public static string GetBranchName(string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        return RunGit(workingDirectory, "branch", "--show-current");
+        return RunGit(workingDirectory, settings, "branch", "--show-current");
     }
 
-    public static void Checkout(string branchName, string workingDirectory = ".")
+    public static void Checkout(string branchName, string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        RunGit(workingDirectory, "checkout", branchName);
+        RunGit(workingDirectory, settings, "checkout", branchName);
     }
 
-    public static void ChangeBranch(string branchName, string workingDirectory = ".")
+    public static void ChangeBranch(string branchName, string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        Checkout(branchName, workingDirectory);
+        Checkout(branchName, workingDirectory, settings);
     }
 }