+++ TJC.GitExtensions/Changes.cs
@@ -2,9 +2,9 @@ namespace TJC.GitExtensions;
 
 public static partial class GitExtensions
 {
-    public static string[] GetChangesList(string workingDirectory = ".")
+    public static string[] GetChangesList(string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        var output = RunGit(workingDirectory, "status", "--short");
+        var output = RunGit(workingDirectory, settings, "status", "--short");
         return string.IsNullOrEmpty(output)
             ? Array.Empty<string>()
             : output.Split(new[] { "\r\n", "\n" }, StringSplitOptions.RemoveEmptyEntries);