+++ TJC.GitExtensions/Fetch.cs
@@ -2,8 +2,8 @@ namespace TJC.GitExtensions;
 
 public static partial class GitExtensions
 {
-    public static void Fetch(string workingDirectory = ".")
+    public static void Fetch(string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        RunGit(workingDirectory, "fetch");
+        RunGit(workingDirectory, settings, "fetch");
     }
 }