+++ TJC.GitExtensions/GetInformation.cs
@@ -2,16 +2,17 @@ namespace TJC.GitExtensions;
 
 public static partial class GitExtensions
 {
-    public static GitRepositoryInfo GetInformation(string workingDirectory = ".")
+    public static GitRepositoryInfo GetInformation(string workingDirectory = ".", GitCommandSettings? settings = null)
     {
         var describe = RunGit(
             workingDirectory,
+            settings,
             "describe",
             "--tags",
             "--always",
             "--long",
             "--abbrev=40");
-        var status = RunGit(workingDirectory, "status", "--short");
+        var status = RunGit(workingDirectory, settings, "status", "--short");
         var description = ParseDescription(describe);
 
         return new GitRepositoryInfo