+++ TJC.GitExtensions/Pull.cs
@@ -2,8 +2,8 @@ namespace TJC.GitExtensions;
 
 public static partial class GitExtensions
 {
-    public static void Pull(string workingDirectory = ".")
+    public static void Pull(string workingDirectory = ".", GitCommandSettings? settings = null)
     {
-        RunGit(workingDirectory, "pull");
+        RunGit(workingDirectory, settings, "pull");
     }
 }